requirements/base.txt
# Project dependencies
# -----------------------

GitPython>=3.1,<3.2
PyGithub>=1.54,<1.58
PyQt5>=5.15,<5.16
pyqt5ac>=1.2,<1.3
python-slugify>=4.0,<8.1
transifex-python
pyyaml>=5.4,<6.1
